refactor: migrate from eslint to biome
