+++ .vscode/settings.json
@@ -1,29 +1,22 @@
 {
-  "files.exclude": {
-    "**/node_modules": true,
-    "**/.DS_Store": true
-  },
-  "search.exclude": {
-    "**/node_modules": true,
-    "**/__snapshots__": true,
-    "build": true,
-    "package-lock.json": true,
-    "yarn.lock": true,
-    ".idea": true,
-    "**/*.svg": true
-  },
   "breadcrumbs.enabled": true,
-  "files.insertFinalNewline": true,
   "editor.bracketPairColorization.enabled": true,
+  "editor.codeActionsOnSave": {
+    "quickfix.biome": "explicit",
+    "source.addMissingImports": "explicit",
+    "source.fixAll": "explicit",
+    "source.organizeImports": "explicit"
+  },
   "editor.copyWithSyntaxHighlighting": false,
-  "editor.defaultFormatter": "esbenp.prettier-vscode",
-  "editor.hover.sticky": true,
+  "editor.defaultFormatter": "biomejs.biome",
+  "editor.formatOnSave": true,
   "editor.guides.bracketPairs": true,
   "editor.guides.indentation": true,
   "editor.guides.highlightActiveIndentation": true,
-  "editor.stickyScroll.enabled": false,
+  "editor.hover.sticky": true,
   "editor.renderControlCharacters": true,
-  "eslint.validate": ["typescript", "typescriptreact"],
+  "editor.stickyScroll.enabled": false,
+  "eslint.enable": false,
   "files.associations": {
     "*.env.local": "env",
     "*.env.development": "env",
@@ -33,10 +26,38 @@
     "*.env.production": "env",
     "*.env.production.local": "env"
   },
+  "files.exclude": {
+    "**/node_modules": true,
+    "**/.DS_Store": true
+  },
+  "files.insertFinalNewline": true,
   "git.mergeEditor": true,
+  // javascript
+  "javascript.suggest.autoImports": true,
+  "javascript.format.enable": false,
+  "javascript.inlayHints.enumMemberValues.enabled": true,
+  "javascript.inlayHints.functionLikeReturnTypes.enabled": true,
+  "javascript.inlayHints.parameterNames.enabled": "all",
+  "javascript.inlayHints.parameterTypes.enabled": true,
+  "javascript.inlayHints.propertyDeclarationTypes.enabled": true,
+  "javascript.inlayHints.variableTypes.enabled": false,
+  "javascript.preferences.importModuleSpecifier": "non-relative",
+  "javascript.referencesCodeLens.enabled": true,
+  "javascript.referencesCodeLens.showOnAllFunctions": false,
   "path-autocomplete.pathMappings": {
     "@src": "${folder}/src",
     "~public": "${folder}/public"
+  },
+  "search.exclude": {
+    "**/node_modules": true,
+    "**/__snapshots__": true,
+    "**/*.svg": true,
+    ".idea": true,
+    "build": true,
+    "bun.lockb": true,
+    "package-lock.json": true,
+    "pnpm-lock.yaml": true,
+    "yarn.lock": true
   },
   // typescript
   "typescript.suggest.autoImports": true,
@@ -51,72 +72,16 @@
   "typescript.preferences.importModuleSpecifier": "non-relative",
   "typescript.referencesCodeLens.enabled": true,
   "typescript.referencesCodeLens.showOnAllFunctions": false,
-  // javascript
-  "javascript.suggest.autoImports": true,
-  "javascript.format.enable": false,
-  "javascript.inlayHints.enumMemberValues.enabled": true,
-  "javascript.inlayHints.functionLikeReturnTypes.enabled": true,
-  "javascript.inlayHints.parameterNames.enabled": "all",
-  "javascript.inlayHints.parameterTypes.enabled": true,
-  "javascript.inlayHints.propertyDeclarationTypes.enabled": true,
-  "javascript.inlayHints.variableTypes.enabled": false,
-  "javascript.preferences.importModuleSpecifier": "non-relative",
-  "javascript.referencesCodeLens.enabled": true,
-  "javascript.referencesCodeLens.showOnAllFunctions": false,
-  "markdown.validate.enabled": true,
   "workbench.editor.closeEmptyGroups": true,
   "workbench.editor.enablePreview": false,
   "workbench.tree.indent": 16,
   "workbench.tree.renderIndentGuides": "always",
-  "[typescript]": {
-    "editor.defaultFormatter": "esbenp.prettier-vscode",
-    "editor.formatOnSave": true,
-    "editor.codeActionsOnSave": {
-      "source.organizeImports": "explicit",
-      "source.addMissingImports": "explicit"
-    }
-  },
-  "[typescriptreact]": {
-    "editor.defaultFormatter": "esbenp.prettier-vscode",
-    "editor.formatOnSave": true,
-    "editor.codeActionsOnSave": {
-      "source.organizeImports": "explicit",
-      "source.addMissingImports": "explicit"
-    }
-  },
-  "[json]": {
-    "editor.defaultFormatter": "esbenp.prettier-vscode",
-    "editor.formatOnSave": true,
-    "editor.codeActionsOnSave": {
-      "source.organizeImports": "explicit"
-    }
-  },
-  "[jsonc]": {
-    "editor.defaultFormatter": "esbenp.prettier-vscode",
-    "editor.formatOnSave": true,
-    "editor.codeActionsOnSave": {
-      "source.organizeImports": "explicit"
-    }
-  },
-  "[css]": {
-    "editor.defaultFormatter": "esbenp.prettier-vscode",
-    "editor.formatOnSave": true,
-    "editor.codeActionsOnSave": {
-      "source.organizeImports": "explicit"
-    }
-  },
   "[scss]": {
     "editor.defaultFormatter": "esbenp.prettier-vscode",
-    "editor.formatOnSave": true,
-    "editor.codeActionsOnSave": {
-      "source.organizeImports": "explicit"
-    }
+    "editor.formatOnSave": true
   },
   "[markdown]": {
     "editor.defaultFormatter": "esbenp.prettier-vscode",
-    "editor.formatOnSave": true,
-    "editor.codeActionsOnSave": {
-      "source.organizeImports": "explicit"
-    }
+    "editor.formatOnSave": true
   }
 }
