+++ package.json
@@ -8,52 +8,47 @@
     "build": "tsc && vite build",
     "preview": "vite preview",
     "test": "vitest",
+    "biome": "biome",
+    "format": "biome format --write",
+    "lint": "biome lint --write",
+    "check": "biome check --write",
     "prepare": "husky install",
     "lint-staged": "lint-staged --allow-empty"
   },
   "dependencies": {
-    "@emotion/react": "^11.11.4",
-    "@emotion/styled": "^11.11.5",
-    "@mui/icons-material": "^5.15.15",
-    "@mui/lab": "^5.0.0-alpha.170",
-    "@mui/material": "^5.15.15",
+    "@emotion/react": "^11.13.3",
+    "@emotion/styled": "^11.13.0",
+    "@mui/icons-material": "^6.1.0",
+    "@mui/lab": "^6.0.0-beta.9",
+    "@mui/material": "^6.1.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^5.0.1",
     "react-router-dom": "^6.22.3"
   },
   "devDependencies": {
+    "@biomejs/biome": "1.9.0",
     "@testing-library/jest-dom": "^6.4.2",
-    "@testing-library/react": "^15.0.0",
+    "@testing-library/react": "16.0.1",
     "@testing-library/user-event": "^14.5.2",
     "@tsconfig/vite-react": "^3.0.2",
     "@types/react": "^18.2.75",
     "@types/react-dom": "^18.2.24",
-    "@types/testing-library__user-event": "^4.2.0",
-    "@typescript-eslint/eslint-plugin": "^7.6.0",
-    "@typescript-eslint/parser": "^7.6.0",
+    "@types/testing-library__user-event": "4.1.1",
     "@vitejs/plugin-react": "^4.2.1",
     "@vitejs/plugin-react-swc": "^3.6.0",
-    "eslint": "^8.56.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-jest-dom": "^5.2.0",
-    "eslint-plugin-react": "^7.34.1",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.6",
-    "eslint-plugin-testing-library": "^6.2.0",
-    "eslint-plugin-unused-imports": "^3.1.0",
     "husky": "^9.0.11",
     "jest-extended": "^4.0.2",
-    "jsdom": "^24.0.0",
+    "jsdom": "25.0.0",
     "lint-staged": "^15.2.2",
-    "npm-check-updates": "^16.14.18",
+    "npm-check-updates": "17.1.1",
     "organize-imports-cli": "^0.10.0",
-    "prettier": "^3.2.5",
+    "prettier": "^3.3.2",
     "sass": "^1.74.1",
     "typescript": "^5.4.5",
-    "vite": "^5.2.8",
-    "vite-plugin-checker": "^0.6.4",
-    "vite-tsconfig-paths": "^4.3.2",
-    "vitest": "^1.4.0"
+    "vite": "5.4.5",
+    "vite-plugin-checker": "^0.8.0",
+    "vite-tsconfig-paths": "5.0.1",
+    "vitest": "2.1.0"
   }
 }
