+++ src/components/layout/ContentHeader/container.tsx
@@ -1,6 +1,6 @@
-import { ContentHeaderPresenter } from "@src/components/layout/ContentHeader";
-import { Layout } from "@src/components/layout/Layout";
-import { ComponentProps } from "react";
+import { ContentHeaderPresenter } from "@src/components/layout/ContentHeader/presenter";
+import type { Layout } from "@src/components/layout/Layout/container";
+import type { ComponentProps } from "react";
 
 type Props = Pick<ComponentProps<typeof Layout>, "pageTitle">;
