+++ src/components/layout/ContentHeader/presenter.tsx
@@ -1,6 +1,6 @@
 import { Card, Typography } from "@mui/material";
-import { Layout } from "@src/components/layout/Layout";
-import { ComponentProps } from "react";
+import type { Layout } from "@src/components/layout/Layout/container";
+import type { ComponentProps } from "react";
 
 type Props = Pick<ComponentProps<typeof Layout>, "pageTitle">;
