+++ src/main.tsx
@@ -1,14 +1,13 @@
-import { createTheme, ThemeProvider } from "@mui/material";
+import { ThemeProvider, createTheme } from "@mui/material";
 import App from "@src/App";
 import "@src/index.scss";
 import { StrictMode } from "react";
-import ReactDOM from "react-dom/client";
+import reactDom from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 
 const theme = createTheme({});
 
-// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-ReactDOM.createRoot(document.getElementById("root")!).render(
+reactDom.createRoot(document.getElementById("root")!).render(
   <StrictMode>
     <BrowserRouter>
       <ThemeProvider theme={theme}>
