--- src/pages/ErrorPage/index.ts
@@ -1,2 +1,0 @@
-export * from "./container";
-export * from "./presenter";
