--- src/pages/HomePage/index.ts
@@ -1,2 +1,0 @@
-export * from "./container";
-export * from "./presenter";
